+++ CHANGELOG.md
@@ -14,6 +14,7 @@ Change Log
 
 - Enable the enhanced runtime protection on macOS 10.14 Mojave also by the MAS version.
 - Make the i-beam cursor more legible in a dark background theme.
+- Sync sidebar width among tabbed windows.
 
 
 ### Fixes